Make it buildable in freestanding environment
In freestanding environment, it should not use features that depend on std.
In normal installation, it should be able to use features that depend on std.
Currently, only the process of converting bit lines to strings depends on std.

This is useful when you want to use only bitwise operations but not string processing from a processor that does not have string processing.